+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ### Added
 
-- New Address Search api client apps/geodata/wfs_clients [PR #6537]
+- New Address Search api client apps/geodata/geosearch_client [PR #6537]
 - Showing Map for Modules where Map is the default view now (#738)
 
 ### Changed