+++ CHANGELOG.md
@@ -7,9 +7,12 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ## Unreleased
 
+## v2509.1
+
 ### Added
 
-- New Address Search api client apps/geodata/wfs_clients [PR #6537]
+- New Address Search api client apps/geodata/geosearch_client [PR #6537]
+- Showing Map for Modules where Map is the default view now (#738)
 
 ### Changed